+++ naze32/src/main/sensors/gyro.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-extern int32_t gyroADC[XYZ_AXIS_COUNT];
+extern float gyroADC[XYZ_AXIS_COUNT];
 
 typedef struct gyroConfig_s {
     uint8_t gyroMovementCalibrationThreshold; // people keep forgetting that moving model while init results in wrong gyro offsets. and then they never reset gyro. so this is now on by default.
